+++ android/app/src/main/java/com/on/turip/ui/travel/detail/DayModel.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.common.model
+package com.on.turip.ui.travel.detail
 
 data class DayModel(
     val day: Int,
