+++ android/app/src/main/java/com/on/turip/ui/travel/detail/webview/TuripWebChromeClient.kt
@@ -1,4 +1,4 @@
-package com.on.turip.ui.search.detail.webview
+package com.on.turip.ui.travel.detail.webview
 
 import android.view.View
 import android.webkit.WebChromeClient
